+++ src/index.ts
@@ -3,6 +3,7 @@ import { error } from "./notify/error";
 import { deploy } from "./notify/deploy";
 import { feedback } from "./notify/feedback";
 import { bug } from "./notify/bug";
+import { send } from "./notify/send";
 
 export * from "./types";
 export * from "./notify/contact";
@@ -10,8 +11,10 @@ export * from "./notify/error";
 export * from "./notify/deploy";
 export * from "./notify/feedback";
 export * from "./notify/bug";
+export * from "./notify/send";
 
 export const formcord = {
+    send,
     contact,
     error,
     deploy,